Compute distribution of OriginTime and WaitTime in billions of years
The idea is that we know OriginTime(Earth) is 13.8 billion years. Assume
Earth is a random civ with MaxAngle=0 (since we do not currently see any
signals from aliens). Suppose Earth is civ X. Then we can scale all the
origin times and waits to billions of years by multiplying by
13.8/OriginTime(X). But we don't know X, so what do we do? For each
choice of X (all the civs with MaxAngle=0), produce the list of
OriginTimes, and combine all of these origin times into one big sorted
list; this big sorted list is the final distribution of OriginTimes (the
same procedure will also generate the distribution of MinWaits). If this
is computationally inefficient, we might sample the choices of X instead
of doing all of them by only considering every Kth option.